move EditListMenu to own file
